+++ spinta/config.py
@@ -203,6 +203,9 @@ CONFIG = {
     # Part of the scope will be replaced with a hash to fit scope_max_length.
     'scope_max_length': 60,
 
+    # If True, then scopes are added to access logs.
+    'scope_log' : True,
+
     # If True, then 'id' property is always included into Action.SEARCH result,
     # even if not explicitly asked.
     'always_show_id': False,