+++ spinta/utils/scopes.py
@@ -3,7 +3,6 @@ import hashlib
 import re
 from typing import Any
 from typing import Dict
-
 scope_re = re.compile(r'[^a-z0-9]+', flags=re.IGNORECASE)