+++ CHANGES.rst
@@ -14,6 +14,10 @@ Improvements:
 
 - Added better support for migrations with nested data types (`#722`_).
 
+- Added scope information to access logs (`#903`_).
+
+  .. _#903: https://github.com/atviriduomenys/spinta/issues/903
+
 Bug fix:
 
 - Resolved ambiguity warning messages (`#895`_).